+++ apps/ledger-live-mobile/src/screens/Swap/LiveApp/hooks/useSwapLiveAppCustomHandlers.ts
@@ -1,28 +1,30 @@
 import { LiveAppManifest } from "@ledgerhq/live-common/platform/types";
 import { WalletAPICustomHandlers } from "@ledgerhq/live-common/wallet-api/types";
+import { useNavigation } from "@react-navigation/native";
 import { useMemo } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { usePTXCustomHandlers } from "~/components/WebPTXPlayer/CustomHandlers";
-import { accountsSelector } from "~/reducers/accounts";
+import { flattenAccountsSelector } from "~/reducers/accounts";
 import { swapCustomHandlers } from "../customHandlers";
-import { useNavigation } from "@react-navigation/native";
 
 export function useSwapLiveAppCustomHandlers(manifest: LiveAppManifest) {
-  const accounts = useSelector(accountsSelector);
+  const accounts = useSelector(flattenAccountsSelector);
   const ptxCustomHandlers = usePTXCustomHandlers(manifest, accounts);
   const dispatch = useDispatch();
   const navigation = useNavigation();
 
-  return useMemo<WalletAPICustomHandlers>(
-    () =>
-      ({
-        ...ptxCustomHandlers,
-        ...swapCustomHandlers({
-          accounts,
-          dispatch,
-          navigation,
-        }),
-      }) as WalletAPICustomHandlers,
-    [ptxCustomHandlers, accounts, dispatch, navigation],
-  );
+  return useMemo<WalletAPICustomHandlers>(() => {
+    // Cast to unknown first to handle type incompatibilities
+    const customHandlers = {
+      ...ptxCustomHandlers,
+      ...swapCustomHandlers({
+        accounts,
+        dispatch,
+        navigation,
+      }),
+    } as unknown;
+
+    // Then cast to the final type
+    return customHandlers as WalletAPICustomHandlers;
+  }, [ptxCustomHandlers, accounts, dispatch, navigation]);
 }