+++ libs/live-dmk/src/transport/DeviceManagementKitTransport.test.ts
@@ -13,8 +13,8 @@ let obs: Subject<DeviceSessionState> = new Subject<DeviceSessionState>();
 let transport: DeviceManagementKitTransport;
 describe("DeviceManagementKitTransport", () => {
   beforeAll(async () => {
-    vi.spyOn(deviceManagementKit, "toggleDeviceSessionRefresher").mockImplementation(vi.fn());
-    vi.spyOn(deviceManagementKit, "listenToKnownDevices").mockImplementation(() => {
+    vi.spyOn(deviceManagementKit, "disableDeviceSessionRefresher").mockImplementation(vi.fn());
+    vi.spyOn(deviceManagementKit, "listenToAvailableDevices").mockImplementation(() => {
       return of<DiscoveredDevice[]>([
         {
           id: `test-123`,