+++ libs/device-react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/device-react
 
+## 0.2.23
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/device-core@0.4.17
+
+## 0.2.23-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/device-core@0.4.17-next.0
+
 ## 0.2.22
 
 ### Patch Changes